src/mcp-server-cvm/pyproject.toml
[project]
name = "mcp-server-cvm"
version = "0.1.12"
description = "MCP Server for Tencent Cloud CVM"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "tencentcloud-sdk-python-cvm==3.0.1356",
    "tencentcloud-sdk-python-vpc==3.0.1356",
    "tencentcloud-sdk-python-monitor==3.0.1356",
    "tencentcloud-sdk-python-common==3.0.1356"
]

[project.scripts]
mcp-server-cvm = "mcp_server_cvm:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_cvm"]
